requirements.txt
fastapi
uvicorn
daytona-sdk
